Extract media and service logic from main index into focused modules
