+++ opennow-stable/src/main/index.ts
@@ -9,22 +9,8 @@ import {
   protocol,
 } from "electron";
 import { fileURLToPath } from "node:url";
-import { dirname, join, resolve } from "node:path";
-import { existsSync, readFileSync, createWriteStream } from "node:fs";
-import {
-  copyFile,
-  mkdir,
-  readdir,
-  readFile,
-  rename,
-  stat,
-  unlink,
-  writeFile,
-  realpath,
-} from "node:fs/promises";
-import * as net from "node:net";
-import { randomUUID, createHash } from "node:crypto";
-import { spawn } from "node:child_process";
+import { dirname, join } from "node:path";
+import { existsSync, readFileSync } from "node:fs";
 import { Buffer } from "node:buffer";
 
 // Keyboard shortcuts reference (matching Rust implementation):
@@ -34,11 +20,7 @@ import { Buffer } from "node:buffer";
 // F8  - Toggle mouse/pointer lock (handled in main process via IPC)
 
 import { IPC_CHANNELS } from "@shared/ipc";
-import {
-  getTrustedVideoPlaybackFileUrl,
-  registerOpenNowMediaProtocol,
-  resolveTrustedOpenNowMediaPath,
-} from "./mediaPaths";
+import { registerOpenNowMediaProtocol } from "./mediaPaths";
 import { initLogCapture, exportLogs } from "@shared/logger";
 import type { NativeStreamerInputPacket } from "@shared/nativeStreamer";
 import { cacheManager } from "./services/cacheManager";
@@ -81,18 +63,6 @@ import type {
   PingResult,
   StreamRegion,
   VideoAccelerationPreference,
-  ScreenshotDeleteRequest,
-  ScreenshotEntry,
-  ScreenshotSaveAsRequest,
-  ScreenshotSaveAsResult,
-  ScreenshotSaveRequest,
-  RecordingEntry,
-  RecordingBeginRequest,
-  RecordingBeginResult,
-  RecordingChunkRequest,
-  RecordingFinishRequest,
-  RecordingAbortRequest,
-  RecordingDeleteRequest,
   MicrophonePermissionResult,
   NativeStreamerStatus,
   ThankYouContributor,
@@ -144,6 +114,13 @@ import {
 } from "./updater";
 import { NativeStreamerManager } from "./nativeStreamer/manager";
 import { getNativeCloudGsyncCapabilities } from "./nativeCloudGsync";
+import { registerMediaIpcHandlers } from "./ipc/mediaHandlers";
+import { fetchWithTimeout, withTimeout } from "./services/requestTimeout";
+import {
+  fetchPrintedWasteQueue,
+  fetchPrintedWasteServerMapping,
+} from "./services/printedWaste";
+import { pingRegions } from "./services/regionPing";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -314,7 +291,6 @@ const MAX_NATIVE_INPUT_PACKET_BYTES = 4096;
 let authService: AuthService;
 let settingsManager: SettingsManager;
 let appUpdater: AppUpdaterController | null = null;
-const SCREENSHOT_LIMIT = 60;
 const EXPLICIT_SHUTDOWN_FORCE_EXIT_DELAY_MS = 2000;
 let isShutdownRequested = false;
 let isShutdownCleanupComplete = false;
@@ -509,393 +485,6 @@ class DiscordStatusMonitor {
 }
 
 const discordMonitor = new DiscordStatusMonitor();
-
-function getScreenshotDirectory(): string {
-  return join(app.getPath("pictures"), "OpenNOW", "Screenshots");
-}
-
-async function ensureScreenshotDirectory(): Promise<string> {
-  const dir = getScreenshotDirectory();
-  await mkdir(dir, { recursive: true });
-  return dir;
-}
-
-function sanitizeTitleForFileName(value: string | undefined): string {
-  const source = (value ?? "").trim().toLowerCase();
-  const compact = source.replace(/[^a-z0-9]+/g, "-").replace(/^-+|-+$/g, "");
-  if (!compact) return "stream";
-  return compact.slice(0, 48);
-}
-
-function dataUrlToBuffer(dataUrl: string): {
-  ext: "png" | "jpg" | "webp";
-  buffer: Buffer;
-} {
-  const match =
-    /^data:image\/(png|jpeg|jpg|webp);base64,([a-z0-9+/=\s]+)$/i.exec(dataUrl);
-  if (!match || !match[1] || !match[2]) {
-    throw new Error("Invalid screenshot payload");
-  }
-
-  const rawExt = match[1].toLowerCase();
-  const ext: "png" | "jpg" | "webp" =
-    rawExt === "jpeg" ? "jpg" : (rawExt as "png" | "jpg" | "webp");
-  const buffer = Buffer.from(match[2].replace(/\s+/g, ""), "base64");
-  if (!buffer.length) {
-    throw new Error("Empty screenshot payload");
-  }
-
-  return { ext, buffer };
-}
-
-function buildScreenshotDataUrl(ext: string, buffer: Buffer): string {
-  const mime =
-    ext === "jpg" ? "image/jpeg" : ext === "webp" ? "image/webp" : "image/png";
-  return `data:${mime};base64,${buffer.toString("base64")}`;
-}
-
-function assertSafeScreenshotId(id: string): void {
-  if (!id || id.includes("/") || id.includes("\\") || id.includes("..")) {
-    throw new Error("Invalid screenshot id");
-  }
-}
-
-async function listScreenshots(): Promise<ScreenshotEntry[]> {
-  const dir = await ensureScreenshotDirectory();
-  const entries = await readdir(dir, { withFileTypes: true });
-  const screenshotFiles = entries
-    .filter((entry) => entry.isFile())
-    .map((entry) => entry.name)
-    .filter((name) => /\.(png|jpg|jpeg|webp)$/i.test(name));
-
-  const loaded = await Promise.all(
-    screenshotFiles.map(async (fileName): Promise<ScreenshotEntry | null> => {
-      const filePath = join(dir, fileName);
-      try {
-        const fileStats = await stat(filePath);
-        const fileBuffer = await readFile(filePath);
-        const extMatch = /\.([^.]+)$/.exec(fileName);
-        const ext = (extMatch?.[1] ?? "png").toLowerCase();
-
-        return {
-          id: fileName,
-          fileName,
-          filePath,
-          createdAtMs: fileStats.birthtimeMs || fileStats.mtimeMs,
-          sizeBytes: fileStats.size,
-          dataUrl: buildScreenshotDataUrl(ext, fileBuffer),
-        };
-      } catch {
-        return null;
-      }
-    }),
-  );
-
-  return loaded
-    .filter((item): item is ScreenshotEntry => item !== null)
-    .sort((a, b) => b.createdAtMs - a.createdAtMs)
-    .slice(0, SCREENSHOT_LIMIT);
-}
-
-async function saveScreenshot(
-  input: ScreenshotSaveRequest,
-): Promise<ScreenshotEntry> {
-  const { ext, buffer } = dataUrlToBuffer(input.dataUrl);
-  const dir = await ensureScreenshotDirectory();
-  const stamp = new Date().toISOString().replace(/[:.]/g, "-");
-  const title = sanitizeTitleForFileName(input.gameTitle);
-  const fileName = `${stamp}-${title}-${Math.random().toString(16).slice(2, 8)}.${ext}`;
-  const filePath = join(dir, fileName);
-
-  await writeFile(filePath, buffer);
-
-  return {
-    id: fileName,
-    fileName,
-    filePath,
-    createdAtMs: Date.now(),
-    sizeBytes: buffer.byteLength,
-    dataUrl: buildScreenshotDataUrl(ext, buffer),
-  };
-}
-
-async function deleteScreenshot(input: ScreenshotDeleteRequest): Promise<void> {
-  assertSafeScreenshotId(input.id);
-  const dir = await ensureScreenshotDirectory();
-  const filePath = join(dir, input.id);
-  await unlink(filePath);
-}
-
-async function saveScreenshotAs(
-  input: ScreenshotSaveAsRequest,
-): Promise<ScreenshotSaveAsResult> {
-  assertSafeScreenshotId(input.id);
-  const dir = await ensureScreenshotDirectory();
-  const sourcePath = join(dir, input.id);
-
-  const saveDialogOptions = {
-    title: "Save Screenshot",
-    defaultPath: join(app.getPath("pictures"), input.id),
-    filters: [
-      { name: "PNG Image", extensions: ["png"] },
-      { name: "JPEG Image", extensions: ["jpg", "jpeg"] },
-      { name: "WebP Image", extensions: ["webp"] },
-      { name: "All Files", extensions: ["*"] },
-    ],
-  };
-  const target =
-    mainWindow && !mainWindow.isDestroyed()
-      ? await dialog.showSaveDialog(mainWindow, saveDialogOptions)
-      : await dialog.showSaveDialog(saveDialogOptions);
-
-  if (target.canceled || !target.filePath) {
-    return { saved: false };
-  }
-
-  await copyFile(sourcePath, target.filePath);
-  return { saved: true, filePath: target.filePath };
-}
-
-// ---------------------------------------------------------------------------
-// Recording helpers
-// ---------------------------------------------------------------------------
-
-const RECORDING_LIMIT = 20;
-
-interface ActiveRecording {
-  writeStream: ReturnType<typeof createWriteStream>;
-  tempPath: string;
-  mimeType: string;
-}
-
-const activeRecordings = new Map<string, ActiveRecording>();
-
-function getRecordingsDirectory(): string {
-  return join(app.getPath("pictures"), "OpenNOW", "Recordings");
-}
-
-function getThumbnailCacheDirectory(): string {
-  return join(app.getPath("userData"), "media-thumbs");
-}
-
-async function ensureThumbnailCacheDirectory(): Promise<string> {
-  const dir = getThumbnailCacheDirectory();
-  await mkdir(dir, { recursive: true });
-  return dir;
-}
-
-function md5(input: string): string {
-  return createHash("md5").update(input).digest("hex");
-}
-
-/** Seconds; null if ffprobe missing or unreadable. */
-async function probeVideoDurationSeconds(
-  sourcePath: string,
-): Promise<number | null> {
-  return new Promise((resolve) => {
-    const args = [
-      "-v",
-      "error",
-      "-show_entries",
-      "format=duration",
-      "-of",
-      "default=noprint_wrappers=1:nokey=1",
-      sourcePath,
-    ];
-    const child = spawn("ffprobe", args, {
-      stdio: ["ignore", "pipe", "ignore"],
-    });
-    let out = "";
-    child.stdout?.on("data", (chunk: Buffer) => {
-      out += chunk.toString();
-    });
-    child.on("error", () => resolve(null));
-    child.on("close", (code) => {
-      if (code !== 0) {
-        resolve(null);
-        return;
-      }
-      const n = Number.parseFloat(out.trim());
-      resolve(Number.isFinite(n) && n > 0 ? n : null);
-    });
-  });
-}
-
-function randomThumbnailSeekSeconds(durationSec: number | null): number {
-  if (durationSec !== null && durationSec > 0.2) {
-    const margin = Math.min(0.35, durationSec * 0.08);
-    const hi = Math.max(durationSec - margin, margin + 0.05);
-    const lo = Math.min(margin, hi * 0.5);
-    return lo + Math.random() * (hi - lo);
-  }
-  return 0.2 + Math.random() * 4.8;
-}
-
-function ffmpegExtractOneFrame(
-  sourcePath: string,
-  outPath: string,
-  seekSec: number,
-): Promise<boolean> {
-  const ss = seekSec.toFixed(3);
-  return new Promise<boolean>((resolve) => {
-    const args = [
-      "-y",
-      "-ss",
-      ss,
-      "-i",
-      sourcePath,
-      "-frames:v",
-      "1",
-      "-q:v",
-      "2",
-      outPath,
-    ];
-    const child = spawn("ffmpeg", args, { stdio: "ignore" });
-    child.on("error", () => resolve(false));
-    child.on("close", (code) => {
-      resolve(code === 0);
-    });
-  });
-}
-
-async function generateVideoThumbnail(
-  sourcePath: string,
-  outPath: string,
-): Promise<boolean> {
-  const durationSec = await probeVideoDurationSeconds(sourcePath);
-  const seekSec = randomThumbnailSeekSeconds(durationSec);
-  if (await ffmpegExtractOneFrame(sourcePath, outPath, seekSec)) return true;
-  if (seekSec > 0.02) return ffmpegExtractOneFrame(sourcePath, outPath, 0);
-  return false;
-}
-
-async function ensureThumbnailForMedia(
-  filePath: string,
-): Promise<string | null> {
-  try {
-    const stats = await stat(filePath);
-    const key = md5(`${filePath}|${stats.mtimeMs}`);
-    const cacheDir = await ensureThumbnailCacheDirectory();
-    const outPath = join(cacheDir, `${key}.jpg`);
-    // If cached, return
-    try {
-      await stat(outPath);
-      return outPath;
-    } catch {
-      // not exists
-    }
-
-    const lower = filePath.toLowerCase();
-    if (
-      lower.endsWith(".mp4") ||
-      lower.endsWith(".webm") ||
-      lower.endsWith(".mkv") ||
-      lower.endsWith(".mov")
-    ) {
-      const ok = await generateVideoThumbnail(filePath, outPath);
-      if (ok) return outPath;
-      // generation failed
-      return null;
-    }
-
-    // For images, copy into cache (no re-encoding)
-    if (
-      lower.endsWith(".png") ||
-      lower.endsWith(".jpg") ||
-      lower.endsWith(".jpeg") ||
-      lower.endsWith(".webp")
-    ) {
-      try {
-        const buf = await readFile(filePath);
-        await writeFile(outPath, buf);
-        return outPath;
-      } catch {
-        return null;
-      }
-    }
-
-    return null;
-  } catch (err) {
-    console.warn("ensureThumbnailForMedia error:", err);
-    return null;
-  }
-}
-
-async function ensureRecordingsDirectory(): Promise<string> {
-  const dir = getRecordingsDirectory();
-  await mkdir(dir, { recursive: true });
-  return dir;
-}
-
-function assertSafeRecordingId(id: string): void {
-  if (!id || id.includes("/") || id.includes("\\") || id.includes("..")) {
-    throw new Error("Invalid recording id");
-  }
-}
-
-function extFromMimeType(mimeType: string): ".mp4" | ".webm" {
-  return mimeType.startsWith("video/mp4") ? ".mp4" : ".webm";
-}
-
-async function listRecordings(): Promise<RecordingEntry[]> {
-  const dir = await ensureRecordingsDirectory();
-  const entries = await readdir(dir, { withFileTypes: true });
-  const webmFiles = entries
-    .filter((e) => e.isFile())
-    .map((e) => e.name)
-    .filter((name) => /\.(mp4|webm)$/i.test(name));
-
-  const loaded = await Promise.all(
-    webmFiles.map(async (fileName): Promise<RecordingEntry | null> => {
-      const filePath = join(dir, fileName);
-      try {
-        const fileStats = await stat(filePath);
-        const stem = fileName.replace(/\.(mp4|webm)$/i, "");
-        const thumbName = `${stem}-thumb.jpg`;
-        const thumbPath = join(dir, thumbName);
-
-        let thumbnailDataUrl: string | undefined;
-        try {
-          const thumbBuf = await readFile(thumbPath);
-          thumbnailDataUrl = `data:image/jpeg;base64,${thumbBuf.toString("base64")}`;
-        } catch {
-          // No thumbnail for this recording — that's fine
-        }
-
-        // Parse durationMs encoded in filename as last numeric segment before extension
-        const durMatch = /-dur(\d+)\.(mp4|webm)$/i.exec(fileName);
-        const durationMs = durMatch ? Number(durMatch[1]) : 0;
-
-        // Parse game title from filename: {stamp}-{title}-{rand}[-dur{ms}].{ext}
-        const titleMatch =
-          /^[^-]+-[^-]+-([^-]+(?:-[^-]+)*?)-[a-f0-9]{6}(?:-dur\d+)?\.(mp4|webm)$/i.exec(
-            fileName,
-          );
-        const gameTitle = titleMatch
-          ? titleMatch[1].replace(/-/g, " ")
-          : undefined;
-
-        return {
-          id: fileName,
-          fileName,
-          filePath,
-          createdAtMs: fileStats.birthtimeMs || fileStats.mtimeMs,
-          sizeBytes: fileStats.size,
-          durationMs,
-          gameTitle,
-          thumbnailDataUrl,
-        };
-      } catch {
-        return null;
-      }
-    }),
-  );
-
-  return loaded
-    .filter((item): item is RecordingEntry => item !== null)
-    .sort((a, b) => b.createdAtMs - a.createdAtMs)
-    .slice(0, RECORDING_LIMIT);
-}
 
 function emitToRenderer(event: MainToRendererSignalingEvent): void {
   if (mainWindow && !mainWindow.isDestroyed()) {
@@ -1475,65 +1064,6 @@ interface GitHubContributorResponse {
   contributions?: number;
   type?: string;
   name?: string | null;
-}
-
-function withTimeout<T>(
-  promise: Promise<T>,
-  timeoutMs: number,
-  label: string,
-): Promise<T> {
-  return new Promise<T>((resolve, reject) => {
-    const timeout = setTimeout(
-      () => reject(new Error(`${label} timed out after ${timeoutMs}ms`)),
-      timeoutMs,
-    );
-    promise.then(
-      (value) => {
-        clearTimeout(timeout);
-        resolve(value);
-      },
-      (error: unknown) => {
-        clearTimeout(timeout);
-        reject(error);
-      },
-    );
-  });
-}
-
-async function fetchWithTimeout(
-  url: string,
-  init: RequestInit,
-  timeoutMs: number,
-  label: string,
-): Promise<Response> {
-  const controller = new AbortController();
-  const timeout = setTimeout(
-    () =>
-      controller.abort(new Error(`${label} timed out after ${timeoutMs}ms`)),
-    timeoutMs,
-  );
-
-  try {
-    return await fetch(url, {
-      ...init,
-      signal: controller.signal,
-    });
-  } catch (error) {
-    if (
-      (error instanceof Error && error.name === "AbortError") ||
-      controller.signal.aborted
-    ) {
-      const reason = controller.signal.reason;
-      const message =
-        reason instanceof Error
-          ? reason.message
-          : `${label} timed out after ${timeoutMs}ms`;
-      throw new Error(message);
-    }
-    throw error;
-  } finally {
-    clearTimeout(timeout);
-  }
 }
 
 function decodeHtmlEntities(value: string): string {
@@ -2587,381 +2117,12 @@ function registerIpcHandlers(): void {
     },
   );
 
-  ipcMain.handle(
-    IPC_CHANNELS.SCREENSHOT_SAVE,
-    async (_event, input: ScreenshotSaveRequest): Promise<ScreenshotEntry> => {
-      return saveScreenshot(input);
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.SCREENSHOT_LIST,
-    async (): Promise<ScreenshotEntry[]> => {
-      return listScreenshots();
-    },
-  );
-
-  // Media: per-game listing (screenshots + recordings). Best-effort title matching.
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_LIST_BY_GAME,
-    async (_event, payload: { gameTitle?: string } = {}) => {
-      const title = (payload?.gameTitle || "").trim().toLowerCase();
-      const screenshots = await listScreenshots();
-      const recordings = await listRecordings();
-
-      const normalize = (s?: string) =>
-        (s || "").replace(/[^a-z0-9]+/gi, "").toLowerCase();
-      const needle = normalize(title);
-
-      const matchedScreens = screenshots.filter((s) => {
-        if (!needle) return true;
-        const candidate = normalize(s.fileName) + normalize(s.filePath || "");
-        return candidate.includes(needle);
-      });
-
-      const matchedRecordings = recordings.filter((r) => {
-        if (!needle) return true;
-        const candidate = normalize(r.gameTitle ?? r.fileName ?? "");
-        return candidate.includes(needle);
-      });
-
-      return {
-        screenshots: matchedScreens,
-        videos: matchedRecordings,
-      };
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.SCREENSHOT_DELETE,
-    async (_event, input: ScreenshotDeleteRequest): Promise<void> => {
-      return deleteScreenshot(input);
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.SCREENSHOT_SAVE_AS,
-    async (
-      _event,
-      input: ScreenshotSaveAsRequest,
-    ): Promise<ScreenshotSaveAsResult> => {
-      return saveScreenshotAs(input);
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_BEGIN,
-    async (
-      _event,
-      input: RecordingBeginRequest,
-    ): Promise<RecordingBeginResult> => {
-      const dir = await ensureRecordingsDirectory();
-      const recordingId = randomUUID();
-      const ext = extFromMimeType(input.mimeType);
-      const tempPath = join(dir, `${recordingId}${ext}.tmp`);
-      const writeStream = createWriteStream(tempPath);
-      activeRecordings.set(recordingId, {
-        writeStream,
-        tempPath,
-        mimeType: input.mimeType,
-      });
-      return { recordingId };
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_CHUNK,
-    async (_event, input: RecordingChunkRequest): Promise<void> => {
-      const rec = activeRecordings.get(input.recordingId);
-      if (!rec) {
-        throw new Error("Unknown recording id");
-      }
-      await new Promise<void>((resolve, reject) => {
-        rec.writeStream.write(Buffer.from(input.chunk), (err) => {
-          if (err) reject(err);
-          else resolve();
-        });
-      });
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_FINISH,
-    async (_event, input: RecordingFinishRequest): Promise<RecordingEntry> => {
-      const rec = activeRecordings.get(input.recordingId);
-      if (!rec) {
-        throw new Error("Unknown recording id");
-      }
-      activeRecordings.delete(input.recordingId);
-
-      await new Promise<void>((resolve, reject) => {
-        rec.writeStream.end((err?: Error | null) => {
-          if (err) reject(err);
-          else resolve();
-        });
-      });
-
-      const dir = getRecordingsDirectory();
-      const stamp = new Date().toISOString().replace(/[:.]/g, "-");
-      const title = sanitizeTitleForFileName(input.gameTitle);
-      const rand = Math.random().toString(16).slice(2, 8);
-      const durSuffix =
-        input.durationMs > 0 ? `-dur${Math.round(input.durationMs)}` : "";
-      const ext = extFromMimeType(rec.mimeType);
-      const fileName = `${stamp}-${title}-${rand}${durSuffix}${ext}`;
-      const finalPath = join(dir, fileName);
-
-      await rename(rec.tempPath, finalPath);
-
-      // Save thumbnail if provided
-      let thumbnailDataUrl: string | undefined;
-      if (input.thumbnailDataUrl) {
-        try {
-          const { buffer } = dataUrlToBuffer(input.thumbnailDataUrl);
-          const stem = fileName.replace(/\.(mp4|webm)$/i, "");
-          const thumbPath = join(dir, `${stem}-thumb.jpg`);
-          await writeFile(thumbPath, buffer);
-          thumbnailDataUrl = input.thumbnailDataUrl;
-        } catch {
-          // Thumbnail save is best-effort — don't fail the recording
-        }
-      }
-
-      // Enforce recording limit: delete oldest entries beyond RECORDING_LIMIT
-      const all = await listRecordings();
-      if (all.length > RECORDING_LIMIT) {
-        const toDelete = all.slice(RECORDING_LIMIT);
-        await Promise.all(
-          toDelete.map(async (entry) => {
-            await unlink(entry.filePath).catch(() => undefined);
-            const stem = entry.fileName.replace(/\.(mp4|webm)$/i, "");
-            await unlink(join(dir, `${stem}-thumb.jpg`)).catch(() => undefined);
-          }),
-        );
-      }
-
-      const fileStats = await stat(finalPath);
-      return {
-        id: fileName,
-        fileName,
-        filePath: finalPath,
-        createdAtMs: Date.now(),
-        sizeBytes: fileStats.size,
-        durationMs: input.durationMs,
-        gameTitle: input.gameTitle,
-        thumbnailDataUrl,
-      };
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_ABORT,
-    async (_event, input: RecordingAbortRequest): Promise<void> => {
-      const rec = activeRecordings.get(input.recordingId);
-      if (!rec) {
-        return;
-      }
-      activeRecordings.delete(input.recordingId);
-      rec.writeStream.destroy();
-      await unlink(rec.tempPath).catch(() => undefined);
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_LIST,
-    async (): Promise<RecordingEntry[]> => {
-      return listRecordings();
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_DELETE,
-    async (_event, input: RecordingDeleteRequest): Promise<void> => {
-      assertSafeRecordingId(input.id);
-      const dir = await ensureRecordingsDirectory();
-      const filePath = join(dir, input.id);
-      await unlink(filePath);
-      const stem = input.id.replace(/\.(mp4|webm)$/i, "");
-      await unlink(join(dir, `${stem}-thumb.jpg`)).catch(() => undefined);
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.RECORDING_SHOW_IN_FOLDER,
-    async (_event, id: string): Promise<void> => {
-      assertSafeRecordingId(id);
-      const dir = await ensureRecordingsDirectory();
-      shell.showItemInFolder(join(dir, id));
-    },
-  );
-
-  // Return a thumbnail data URL for a given media file path (images or companion thumbs for videos).
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_THUMBNAIL,
-    async (_event, payload: { filePath: string }): Promise<string | null> => {
-      const rawFp = payload?.filePath;
-      if (typeof rawFp !== "string") return null;
-      try {
-        const fpReal = await resolveTrustedOpenNowMediaPath(rawFp);
-        if (!fpReal) return null;
-
-        const lower = fpReal.toLowerCase();
-        if (
-          lower.endsWith(".png") ||
-          lower.endsWith(".jpg") ||
-          lower.endsWith(".jpeg") ||
-          lower.endsWith(".webp")
-        ) {
-          const buf = await readFile(fpReal);
-          const extMatch = /\.([^.]+)$/.exec(fpReal);
-          const ext = (extMatch?.[1] || "png").toLowerCase();
-          const mime =
-            ext === "jpg" || ext === "jpeg"
-              ? "image/jpeg"
-              : ext === "webp"
-                ? "image/webp"
-                : "image/png";
-          return `data:${mime};base64,${buf.toString("base64")}`;
-        }
-
-        if (
-          lower.endsWith(".mp4") ||
-          lower.endsWith(".webm") ||
-          lower.endsWith(".mkv") ||
-          lower.endsWith(".mov")
-        ) {
-          // Prefer an existing companion thumb next to the video
-          const stem = fpReal.replace(/\.(mp4|webm|mkv|mov)$/i, "");
-          const thumbPath = `${stem}-thumb.jpg`;
-          try {
-            const b = await readFile(thumbPath);
-            return `data:image/jpeg;base64,${b.toString("base64")}`;
-          } catch {
-            // Try generating a cached thumbnail via ffmpeg
-          }
-
-          const gen = await ensureThumbnailForMedia(fpReal);
-          if (gen) {
-            try {
-              const b2 = await readFile(gen);
-              return `data:image/jpeg;base64,${b2.toString("base64")}`;
-            } catch {
-              return null;
-            }
-          }
-          return null;
-        }
-
-        return null;
-      } catch (err) {
-        console.warn("MEDIA_THUMBNAIL error:", err);
-        return null;
-      }
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_SHOW_IN_FOLDER,
-    async (_event, payload: { filePath: string }): Promise<void> => {
-      const rawFp = payload?.filePath;
-      if (typeof rawFp !== "string") return;
-      try {
-        const fpReal = await resolveTrustedOpenNowMediaPath(rawFp);
-        if (!fpReal) return;
-        shell.showItemInFolder(fpReal);
-      } catch {
-        return;
-      }
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_PLAYBACK_URL,
-    async (_event, payload: { filePath: string }): Promise<string | null> => {
-      const rawFp = payload?.filePath;
-      if (typeof rawFp !== "string") return null;
-      try {
-        return await getTrustedVideoPlaybackFileUrl(rawFp);
-      } catch (err) {
-        console.warn("MEDIA_PLAYBACK_URL error:", err);
-        return null;
-      }
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_DELETE_FILE,
-    async (_event, payload: { filePath: string }): Promise<{ ok: boolean }> => {
-      const rawFp = payload?.filePath;
-      if (typeof rawFp !== "string") return { ok: false };
-      try {
-        const fpReal = await resolveTrustedOpenNowMediaPath(rawFp);
-        if (!fpReal) return { ok: false };
-        let st: Awaited<ReturnType<typeof stat>>;
-        try {
-          st = await stat(fpReal);
-        } catch {
-          return { ok: false };
-        }
-        const key = md5(`${fpReal}|${st.mtimeMs}`);
-        const cacheDir = await ensureThumbnailCacheDirectory();
-        await unlink(join(cacheDir, `${key}.jpg`)).catch(() => undefined);
-        const stem = fpReal.replace(
-          /\.(mp4|webm|mkv|mov|png|jpg|jpeg|webp)$/i,
-          "",
-        );
-        await unlink(`${stem}-thumb.jpg`).catch(() => undefined);
-        await unlink(fpReal);
-        return { ok: true };
-      } catch (err) {
-        console.warn("MEDIA_DELETE_FILE error:", err);
-        return { ok: false };
-      }
-    },
-  );
-
-  ipcMain.handle(
-    IPC_CHANNELS.MEDIA_REGEN_THUMBNAIL,
-    async (
-      _event,
-      payload: { filePath: string },
-    ): Promise<{ ok: boolean; thumbnailDataUrl: string | null }> => {
-      const rawFp = payload?.filePath;
-      if (typeof rawFp !== "string")
-        return { ok: false, thumbnailDataUrl: null };
-      try {
-        const fpReal = await resolveTrustedOpenNowMediaPath(rawFp);
-        if (!fpReal) return { ok: false, thumbnailDataUrl: null };
-        const st = await stat(fpReal);
-        const key = md5(`${fpReal}|${st.mtimeMs}`);
-        const cacheDir = await ensureThumbnailCacheDirectory();
-        await unlink(join(cacheDir, `${key}.jpg`)).catch(() => undefined);
-        if (/\.(mp4|webm|mkv|mov)$/i.test(fpReal)) {
-          const videoStem = fpReal.replace(/\.(mp4|webm|mkv|mov)$/i, "");
-          await unlink(`${videoStem}-thumb.jpg`).catch(() => undefined);
-        }
-        const genPath = await ensureThumbnailForMedia(fpReal);
-        if (!genPath) return { ok: false, thumbnailDataUrl: null };
-
-        if (/\.(mp4|webm|mkv|mov)$/i.test(fpReal)) {
-          const videoStem = fpReal.replace(/\.(mp4|webm|mkv|mov)$/i, "");
-          const sidecar = `${videoStem}-thumb.jpg`;
-          await copyFile(genPath, sidecar).catch((err) => {
-            console.warn("MEDIA_REGEN_THUMBNAIL sidecar copy:", err);
-          });
-        }
-
-        const b = await readFile(genPath);
-        return {
-          ok: true,
-          thumbnailDataUrl: `data:image/jpeg;base64,${b.toString("base64")}`,
-        };
-      } catch (err) {
-        console.warn("MEDIA_REGEN_THUMBNAIL error:", err);
-        return { ok: false, thumbnailDataUrl: null };
-      }
-    },
-  );
+  registerMediaIpcHandlers({
+    ipcMain,
+    dialog,
+    shell,
+    getMainWindow: () => mainWindow,
+  });
 
   ipcMain.handle(IPC_CHANNELS.CACHE_REFRESH_MANUAL, async (): Promise<void> => {
     await refreshScheduler.manualRefresh();
@@ -2979,273 +2140,20 @@ function registerIpcHandlers(): void {
     },
   );
 
-  // TCP-based ping function - more accurate than HTTP as it only measures connection time
-  async function tcpPing(
-    hostname: string,
-    port: number,
-    timeoutMs: number = 3000,
-  ): Promise<number | null> {
-    return new Promise((resolve) => {
-      const startTime = Date.now();
-      const socket = new net.Socket();
-
-      socket.setTimeout(timeoutMs);
-
-      socket.once("connect", () => {
-        const pingMs = Date.now() - startTime;
-        socket.destroy();
-        resolve(pingMs);
-      });
-
-      socket.once("timeout", () => {
-        socket.destroy();
-        resolve(null);
-      });
-
-      socket.once("error", () => {
-        socket.destroy();
-        resolve(null);
-      });
-
-      socket.connect(port, hostname);
-    });
-  }
-
-  // Ping regions IPC handler - uses TCP connection timing for accurate latency measurement
-  // Runs 3 tests and averages the results
   ipcMain.handle(
     IPC_CHANNELS.PING_REGIONS,
     async (_event, regions: StreamRegion[]): Promise<PingResult[]> => {
-      const pingPromises = regions.map(async (region) => {
-        try {
-          const url = new URL(region.url);
-          const hostname = url.hostname;
-          const port = url.protocol === "https:" ? 443 : 80;
-
-          const validPings: number[] = [];
-
-          // Warm-up ping (result discarded) to prime the TCP path before measuring.
-          // The first cold-start connect includes DNS resolution and TCP SYN overhead
-          // which inflates subsequent measurements if not accounted for.
-          await tcpPing(hostname, port, 3000);
-
-          // Run 3 measured ping tests with a brief delay between each to allow
-          // the previous socket to fully close before opening the next connection.
-          for (let i = 0; i < 3; i++) {
-            if (i > 0) {
-              await new Promise<void>((resolve) => setTimeout(resolve, 100));
-            }
-            const pingMs = await tcpPing(hostname, port, 3000);
-            if (pingMs !== null) {
-              validPings.push(pingMs);
-            }
-          }
-
-          // Calculate average of successful pings
-          if (validPings.length > 0) {
-            const avgPing = Math.round(
-              validPings.reduce((a, b) => a + b, 0) / validPings.length,
-            );
-            return { url: region.url, pingMs: avgPing };
-          } else {
-            return {
-              url: region.url,
-              pingMs: null,
-              error: "All ping tests failed",
-            };
-          }
-        } catch {
-          return {
-            url: region.url,
-            pingMs: null,
-            error: "Invalid URL",
-          };
-        }
-      });
-
-      return Promise.all(pingPromises);
+      return pingRegions(regions);
     },
   );
 
   // PrintedWaste queue API — fetched from main process so User-Agent can be set
   ipcMain.handle(IPC_CHANNELS.PRINTEDWASTE_QUEUE_FETCH, async () => {
-    const PRINTEDWASTE_QUEUE_TIMEOUT_MS = 7000;
-    const version = app.getVersion();
-    const response = await fetchWithTimeout(
-      "https://api.printedwaste.com/gfn/queue/",
-      {
-        headers: {
-          "User-Agent": `opennow/${version}`,
-          Accept: "application/json",
-        },
-      },
-      PRINTEDWASTE_QUEUE_TIMEOUT_MS,
-      "PrintedWaste queue request",
-    );
-    if (!response.ok) {
-      throw new Error(`PrintedWaste API returned HTTP ${response.status}`);
-    }
-
-    const body = await withTimeout(
-      response.json() as Promise<unknown>,
-      PRINTEDWASTE_QUEUE_TIMEOUT_MS,
-      "PrintedWaste queue response parse",
-    );
-    if (!body || typeof body !== "object" || Array.isArray(body)) {
-      throw new Error("PrintedWaste API response was not an object");
-    }
-
-    const apiBody = body as { status?: unknown; data?: unknown };
-    if (typeof apiBody.status !== "boolean") {
-      throw new Error("PrintedWaste API response missing boolean status");
-    }
-    if (!apiBody.status) {
-      throw new Error("PrintedWaste API returned status:false");
-    }
-    if (
-      !apiBody.data ||
-      typeof apiBody.data !== "object" ||
-      Array.isArray(apiBody.data)
-    ) {
-      throw new Error("PrintedWaste API response missing data object");
-    }
-
-    const normalizedData: Record<
-      string,
-      {
-        QueuePosition: number;
-        "Last Updated": number;
-        Region: string;
-        eta?: number;
-      }
-    > = {};
-    for (const [zoneId, rawZone] of Object.entries(
-      apiBody.data as Record<string, unknown>,
-    )) {
-      if (!rawZone || typeof rawZone !== "object" || Array.isArray(rawZone)) {
-        continue;
-      }
-      const zone = rawZone as Record<string, unknown>;
-      const queuePosition = zone.QueuePosition;
-      const lastUpdated = zone["Last Updated"];
-      const region = zone.Region;
-      const eta = zone.eta;
-
-      if (
-        typeof queuePosition !== "number" ||
-        !Number.isFinite(queuePosition)
-      ) {
-        continue;
-      }
-      if (typeof lastUpdated !== "number" || !Number.isFinite(lastUpdated)) {
-        continue;
-      }
-      if (typeof region !== "string" || region.length === 0) {
-        continue;
-      }
-      if (
-        eta !== undefined &&
-        (typeof eta !== "number" || !Number.isFinite(eta))
-      ) {
-        continue;
-      }
-
-      normalizedData[zoneId] = {
-        QueuePosition: queuePosition,
-        "Last Updated": lastUpdated,
-        Region: region,
-        ...(typeof eta === "number" ? { eta } : {}),
-      };
-    }
-
-    if (Object.keys(normalizedData).length === 0) {
-      throw new Error("PrintedWaste API returned no valid zones");
-    }
-    return normalizedData;
+    return fetchPrintedWasteQueue(app.getVersion());
   });
 
   ipcMain.handle(IPC_CHANNELS.PRINTEDWASTE_SERVER_MAPPING_FETCH, async () => {
-    const PRINTEDWASTE_MAPPING_TIMEOUT_MS = 7000;
-    const version = app.getVersion();
-    const response = await fetchWithTimeout(
-      "https://remote.printedwaste.com/config/GFN_SERVERID_TO_REGION_MAPPING",
-      {
-        headers: {
-          "User-Agent": `opennow/${version}`,
-          Accept: "application/json",
-        },
-      },
-      PRINTEDWASTE_MAPPING_TIMEOUT_MS,
-      "PrintedWaste server mapping request",
-    );
-    if (!response.ok) {
-      throw new Error(
-        `PrintedWaste server mapping returned HTTP ${response.status}`,
-      );
-    }
-
-    const body = await withTimeout(
-      response.json() as Promise<unknown>,
-      PRINTEDWASTE_MAPPING_TIMEOUT_MS,
-      "PrintedWaste server mapping response parse",
-    );
-    if (!body || typeof body !== "object" || Array.isArray(body)) {
-      throw new Error("PrintedWaste server mapping response was not an object");
-    }
-
-    const apiBody = body as { status?: unknown; data?: unknown };
-    if (typeof apiBody.status !== "boolean") {
-      throw new Error(
-        "PrintedWaste server mapping response missing boolean status",
-      );
-    }
-    if (!apiBody.status) {
-      throw new Error("PrintedWaste server mapping returned status:false");
-    }
-    if (
-      !apiBody.data ||
-      typeof apiBody.data !== "object" ||
-      Array.isArray(apiBody.data)
-    ) {
-      throw new Error(
-        "PrintedWaste server mapping response missing data object",
-      );
-    }
-
-    const normalizedData: Record<
-      string,
-      {
-        title?: string;
-        region?: string;
-        is4080Server?: boolean;
-        is5080Server?: boolean;
-        nuked?: boolean;
-      }
-    > = {};
-
-    for (const [zoneId, rawZone] of Object.entries(
-      apiBody.data as Record<string, unknown>,
-    )) {
-      if (!rawZone || typeof rawZone !== "object" || Array.isArray(rawZone)) {
-        continue;
-      }
-      const zone = rawZone as Record<string, unknown>;
-      const title = zone.title;
-      const region = zone.region;
-      const is4080Server = zone.is4080Server;
-      const is5080Server = zone.is5080Server;
-      const nuked = zone.nuked;
-
-      normalizedData[zoneId] = {
-        ...(typeof title === "string" ? { title } : {}),
-        ...(typeof region === "string" ? { region } : {}),
-        ...(typeof is4080Server === "boolean" ? { is4080Server } : {}),
-        ...(typeof is5080Server === "boolean" ? { is5080Server } : {}),
-        ...(typeof nuked === "boolean" ? { nuked } : {}),
-      };
-    }
-
-    return normalizedData;
+    return fetchPrintedWasteServerMapping(app.getVersion());
   });
 
   // Save window size when it changes
